got 4 bit quant working on larger model
outliers in the weights caused me to have to manually fine tune the scale and zero points which is not ideal. Previously tried getting the 4 bit quant working on a model that could actually fit into the file but a single weight outlier would completely destroy the entire layer. This is what prompted trying 4 bit quant on the larger model however it still struggled in the final layers with outliers/ 4 bits not being enough so I eventually overcame the problem by just 8 bit quanting the final layer. This is only a tiny fraction of the weights so the extra storage cost is well worth it in actually having none random results. Looking back doing the same on the smaller model which likely be the smartest thing to do if not break away from post training quant and try QAT or some entirely different technique like binary quant.